Extract tests into a deep file hierarchy
This will help to more accurately reflect which tests change based on
which files change, and also makes the files overall shorter.
