+++ src/lib.rs
@@ -243,197 +243,4 @@ impl PartialOrd for NetAddr {
 }
 
 #[cfg(test)]
-mod tests {
-	mod netaddr {
-		use crate::NetAddr;
-
-		use std::net::{IpAddr, Ipv4Addr, Ipv6Addr};
-
-		#[test]
-		fn is_send() {
-			fn assert_send<T: Send>() {}
-			assert_send::<NetAddr>();
-		}
-
-		#[test]
-		fn is_sync() {
-			fn assert_sync<T: Sync>() {}
-			assert_sync::<NetAddr>();
-		}
-
-		mod contains {
-			use super::*;
-
-			#[test]
-			fn v4_seems_correct() {
-				let net: NetAddr = "127.0.0.1/8".parse().unwrap();
-				assert!(net.contains(&IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1))));
-				assert!(net.contains(&IpAddr::V4(Ipv4Addr::new(127, 127, 255, 1))));
-				assert!(!net.contains(&IpAddr::V4(Ipv4Addr::new(64, 0, 0, 0))));
-			}
-		}
-
-		#[test]
-		fn broadcast_v4_seems_correct() {
-			let net: NetAddr = "127.0.0.1/8".parse().unwrap();
-			assert_eq!(
-				net.broadcast().unwrap(),
-				IpAddr::V4(Ipv4Addr::new(127, 255, 255, 255))
-			);
-
-			let net: NetAddr = "192.168.69.25/29".parse().unwrap();
-			assert_eq!(
-				net.broadcast().unwrap(),
-				IpAddr::V4(Ipv4Addr::new(192, 168, 69, 31))
-			);
-
-			let net: NetAddr = "192.168.128.127/32".parse().unwrap();
-			assert_eq!(
-				net.broadcast().unwrap(),
-				IpAddr::V4(Ipv4Addr::new(192, 168, 128, 127))
-			);
-		}
-
-		#[test]
-		fn broadcast_v6_returns_none() {
-			let net: NetAddr = "fe80::1/64".parse().unwrap();
-			assert_eq!(net.broadcast(), None);
-		}
-
-		#[test]
-		fn cmp_v4_different_networks() {
-			let a: NetAddr = "1.0.0.0/8".parse().unwrap();
-			let b: NetAddr = "2.0.0.0/8".parse().unwrap();
-
-			assert_eq!(a.partial_cmp(&b), Some(std::cmp::Ordering::Less))
-		}
-
-		#[test]
-		fn cmp_v4_different_netmasks() {
-			let a: NetAddr = "1.0.0.0/7".parse().unwrap();
-			let b: NetAddr = "1.0.0.0/8".parse().unwrap();
-
-			assert_eq!(a.partial_cmp(&b), Some(std::cmp::Ordering::Less))
-		}
-
-		#[test]
-		fn cmp_v4_different() {
-			let a: NetAddr = "1.0.0.0/8".parse().unwrap();
-			let b: NetAddr = "0.0.0.0/24".parse().unwrap();
-
-			assert_eq!(a.partial_cmp(&b), Some(std::cmp::Ordering::Greater))
-		}
-
-		#[test]
-		fn cmp_v4_equal() {
-			let a: NetAddr = "1.0.0.0/8".parse().unwrap();
-			let b: NetAddr = "1.0.0.0/8".parse().unwrap();
-
-			assert_eq!(a.partial_cmp(&b), Some(std::cmp::Ordering::Equal))
-		}
-
-		mod from {
-			use super::*;
-
-			mod ipaddr {
-				use super::*;
-
-				#[test]
-				fn v4_returns_full_netmask() {
-					let addr: IpAddr = "192.0.2.42".parse().unwrap();
-					assert_eq!(
-						NetAddr::from(addr),
-						NetAddr {
-							network: addr,
-							netmask: IpAddr::V4(Ipv4Addr::from(0xff_ff_ff_ff))
-						}
-					);
-				}
-
-				#[test]
-				fn v6_returns_full_netmask() {
-					let addr: IpAddr = "2001:db8:dead:beef::42".parse().unwrap();
-					assert_eq!(
-						NetAddr::from(addr),
-						NetAddr {
-							network: addr,
-							netmask: IpAddr::V6(Ipv6Addr::from(0xffff_ffff_ffff_ffff_ffff_ffff_ffff_ffff))
-						}
-					);
-				}
-			}
-		}
-
-		mod parse {
-			use super::*;
-
-			#[test]
-			fn invalid_is_safe() {
-				let _: Result<NetAddr, _> = "zoop".parse::<NetAddr>();
-			}
-
-			#[test]
-			fn v4_correct_network() {
-				let net: NetAddr = "192.0.2.0/32".parse().unwrap();
-				assert_eq!(net.netmask, IpAddr::V4(Ipv4Addr::new(255, 255, 255, 255)));
-				assert_eq!(net.network, IpAddr::V4(Ipv4Addr::new(192, 0, 2, 0)));
-			}
-
-			#[test]
-			fn v4_localhost() {
-				let net: NetAddr = "127.0.0.1/8".parse().unwrap();
-				assert_eq!(net.netmask, IpAddr::V4(Ipv4Addr::new(255, 0, 0, 0)));
-				assert_eq!(net.network, IpAddr::V4(Ipv4Addr::new(127, 0, 0, 0)));
-			}
-
-			#[test]
-			fn v4_cidr_22() {
-				let net: NetAddr = "192.168.16.1/22".parse().unwrap();
-				assert_eq!(net.netmask, IpAddr::V4(Ipv4Addr::new(255, 255, 252, 0)));
-				assert_eq!(net.network, IpAddr::V4(Ipv4Addr::new(192, 168, 16, 0)));
-			}
-
-			#[test]
-			fn v4_extended_localhost() {
-				let net: NetAddr = "127.0.0.1 255.0.0.0".parse().unwrap();
-				assert_eq!(net.netmask, IpAddr::V4(Ipv4Addr::new(255, 0, 0, 0)));
-				assert_eq!(net.network, IpAddr::V4(Ipv4Addr::new(127, 0, 0, 0)));
-			}
-
-			#[test]
-			fn v4_slashed_localhost() {
-				let net: NetAddr = "127.0.0.1/255.0.0.0".parse().unwrap();
-				assert_eq!(net.netmask, IpAddr::V4(Ipv4Addr::new(255, 0, 0, 0)));
-				assert_eq!(net.network, IpAddr::V4(Ipv4Addr::new(127, 0, 0, 0)));
-			}
-
-			#[test]
-			fn v6_cidr_8() {
-				let net: NetAddr = "ff02::1/8".parse().unwrap();
-				assert_eq!(
-					net.netmask,
-					IpAddr::V6(Ipv6Addr::new(0xff00, 0, 0, 0, 0, 0, 0, 0x0000))
-				);
-				assert_eq!(
-					net.network,
-					IpAddr::V6(Ipv6Addr::new(0xff00, 0, 0, 0, 0, 0, 0, 0x0000))
-				);
-			}
-
-			#[test]
-			fn v6_cidr_128() {
-				let net: NetAddr = "ff02::1/128".parse().unwrap();
-				assert_eq!(
-					net.netmask,
-					IpAddr::V6(Ipv6Addr::new(
-						0xffff, 0xffff, 0xffff, 0xffff, 0xffff, 0xffff, 0xffff, 0xffff
-					))
-				);
-				assert_eq!(
-					net.network,
-					IpAddr::V6(Ipv6Addr::new(0xff02, 0, 0, 0, 0, 0, 0, 0x0001))
-				);
-			}
-		}
-	}
-}
+mod tests;
