+++ src/tests.rs
@@ -1,0 +1,2 @@
+#[cfg(test)]
+mod netaddr;
