+++ src/tests/netaddr/from.rs
@@ -1,0 +1,3 @@
+use super::*;
+
+#[cfg(test)] mod ipaddr;
